+++ app/templates/books/book/pages.hbs
@@ -1,14 +1,10 @@
 <div id="pageButtons">
   <span class="flexSpace">
-    {{#paper-button data-test-clear-button raised=true onClick=(action 'clearText')}}{{t 'clear.text'}}{{/paper-button}}
-    {{#paper-button data-test-highlight-button raised=true onClick=(action 'highlight')}}{{t 'highlight.text'}}{{/paper-button}}
   </span>
 </div>
 <div id="pageInputs">
   <span class="flexSpace">
-    {{paper-input data-test-family-input label=(t 'current.font') value=family onChange=(action "mutFamily") icon="font_download" id="fontFamily"}}
-    {{paper-input data-test-size-input label=(t 'font.size') value=size onChange=(action "mutSize") icon="format_size" id="fontSize"}}
-    {{paper-input data-test-highlight-input label=(t 'highlighted.text') value=highlightText onChange=(action (mut highlightText)) icon="highlight"}}
+    {{page-features}}
   </span>
 </div>
 
@@ -29,7 +25,7 @@
   </span>
 </span>
 <span id='page-numbers'>
-  {{#each model.pages as |page index| }}
+  {{#each model.pages as |page index|}}
     {{#link-to 'books.book.pages.page' page}}<span data-test-page-number>{{increment index}}</span>{{/link-to}}
   {{/each}}
 </span>
