+++ app/templates/components/page-editor.hbs
@@ -1,2 +1,2 @@
-<span contenteditable="true" id="textArea" style={{textAreaStyle}} data-test-text-area>{{page.content}}</span>
+<span contenteditable="true" id="textArea" style={{page-features.textAreaStyle}} data-test-text-area>{{page.content}}</span>
 {{yield}}
