+++ app/templates/components/page-features.hbs
@@ -1,0 +1,6 @@
+{{paper-input data-test-family-input label=(t 'current.font') value=family onChange=(action "mutFamily") icon="font_download" id="fontFamily"}}
+{{paper-input data-test-size-input label=(t 'font.size') value=size onChange=(action "mutSize") icon="format_size" id="fontSize"}}
+{{paper-input data-test-highlight-input label=(t 'highlighted.text') value=highlightText onChange=(action (mut highlightText)) icon="highlight"}}
+{{#paper-button data-test-clear-button raised=true onClick=(action 'clearText')}}{{t 'clear.text'}}{{/paper-button}}
+{{#paper-button data-test-highlight-button raised=true onClick=(action 'highlight')}}{{t 'highlight.text'}}{{/paper-button}}
+{{yield}}
