+++ mirage/factories/page.js
@@ -1,13 +1,12 @@
-import { Factory, faker } from 'ember-cli-mirage';
+import { Factory } from 'ember-cli-mirage';
 
 export default Factory.extend({
 
   content() {
-    return "";
+    return '';
   },
 
   bookId() {
     return 1;
   }
-
 });
